fix potato throw
Potato now gets thrown in direction player is facing even when the player isn't moving (default is down if no movement has taken place). There is a slight issue where if a player moves diagonally, stops, and throws, the potato goes only in the direction of one component of the diagonal (whichever one was let go of a fraction of a second after the first).