Fixed Issue #1
Why this error happened:
When Player 1 dashes and pushes Player 2, Player 2 inherits a higher velocity from the collision.
As soon as Player 1 stops dashing, their movement speed returns to normal, and Player 2's velocity is immediately returns to normal velocity as the current velocity is overridden by ApplyMovementSpeed(), which ignores any momentum from the push.

How I fixed:
I created a variable called pushedVelocity which stores the velocity during pushing and assigned that velocity into the update function.
I also created a boolean variable called isPusher to determine if the object is being pushed or not. Only the object that is being pushed has the effect of gradually slow down.